+++ .circleci/config.yml
@@ -11,36 +11,3 @@ jobs:
       - run: make
       - run: make test
       # - run: make lint
-
-
-
-# machine:
-#   services:
-#     - docker
-  # environment:
-  #   GOPATH: /home/ubuntu
-  #   SRCDIR: /home/ubuntu/scope-plugin
-  #   PATH: $PATH:$HOME/.local/bin
-
-# dependencies:
-#   cache_directories:
-#     - "~/docker"
-#   override:
-#     - echo "no dependencies"
-
-# test:
-#   override:
-#     - cd $SRCDIR && make .iops.uptodate && docker tag weaveworksplugins/scope-iops pradeepkumar95/scope-plugin:$(./tools/image-tag)
-#     - docker tag weaveworksplugins/scope-iops pradeepkumar95/scope-plugin:latest
-
-# deployment:
-#   stag:
-#     branch: staging
-#     commands:
-#       - docker login -e $DOCKER_EMAIL -u $DOCKER_USER -p $DOCKER_PASS
-#       - docker push pradeepkumar95/scope-plugin:$(./tools/image-tag)
-#   pro:
-#     branch: master
-#     commands:
-#       - docker login -e $DOCKER_EMAIL -u $DOCKER_USER -p $DOCKER_PASS
-#       - docker push pradeepkumar95/scope-plugin:latest
