+++ Dockerfile
@@ -1,9 +1,9 @@
 FROM ubuntu 
-ENV CORTEXAGENT=""
-RUN apt-get -y update   
+
+RUN apt-get -y update
+
 RUN apt-get install -y curl
-COPY ./iops /usr/bin/iops
-LABEL works.weave.role=system
-ENTRYPOINT ["/usr/bin/iops"]
 
+COPY ./scope-plugin /usr/bin/scope-plugin
 
+ENTRYPOINT ["/usr/bin/scope-plugin"]
