--- tools/.gitignore
@@ -1,7 +1,0 @@
-cover/cover
-socks/proxy
-socks/image.tar
-runner/runner
-cmd/wcloud/wcloud
-*.pyc
-*~
