--- tools/files-with-type
@@ -1,11 +1,0 @@
-#!/usr/bin/env bash
-#
-# Find all files with a given MIME type.
-#
-# e.g.
-#     $ files-with-type text/x-shellscript k8s infra
-
-mime_type=$1
-shift
-
-git ls-files "$@" | grep -vE '^vendor/' | xargs file --mime-type | grep "${mime_type}" | sed -e 's/:.*$//'
