--- tools/scheduler/.gitignore
@@ -1,1 +1,0 @@
-lib
