--- tools/scheduler/README.md
@@ -1,6 +1,0 @@
-To upload newer version:
-
-```
-pip install -r requirements.txt -t lib
-appcfg.py update .
-```
