--- tools/scheduler/app.yaml
@@ -1,15 +1,0 @@
-application: positive-cocoa-90213
-version: 1
-runtime: python27
-api_version: 1
-threadsafe: true
-
-handlers:
-- url: .*
-  script: main.app
-
-libraries:
-- name: webapp2
-  version: latest
-- name: ssl
-  version: latest
