--- tools/scheduler/appengine_config.py
@@ -1,3 +1,0 @@
-from google.appengine.ext import vendor
-
-vendor.add('lib')
