--- tools/scheduler/cron.yaml
@@ -1,4 +1,0 @@
-cron:
-- description: periodic gc
-  url: /tasks/gc
-  schedule: every 5 minutes
