--- tools/scheduler/requirements.txt
@@ -1,2 +1,0 @@
-flask
-google-api-python-client
