--- tools/shell-lint
@@ -1,13 +1,0 @@
-#!/usr/bin/env bash
-#
-# Lint all shell files in given directories with `shellcheck`.
-#
-# e.g.
-#     $ shell-lint infra k8s
-#
-# Depends on:
-# - shellcheck
-# - files-with-type
-# - file >= 5.22
-
-"$(dirname "${BASH_SOURCE[0]}")/files-with-type" text/x-shellscript "$@" | xargs --no-run-if-empty shellcheck
