--- tools/socks/Dockerfile
@@ -1,7 +1,0 @@
-FROM gliderlabs/alpine
-MAINTAINER Weaveworks Inc <help@weave.works>
-WORKDIR /
-COPY proxy /
-EXPOSE 8000
-EXPOSE 8080
-ENTRYPOINT ["/proxy"]
